requirements.txt
aiohttp==2.3.1
pytest-aiohttp==0.1.3
